refactor(studio): extract MetricTrend from MetricTrendPanel
Addresses review feedback: pull the chart out into its own component
rather than gating the panel chrome behind a `chartOnly` prop.

`MetricTrend` renders the value, delta, evaluator pills, and trendline,
and owns the selected-series state. `MetricTrendPanel` is now just the
titled panel around it, passing its title down as the pill group's
accessible name and the bleed margins as a className. The panel keeps
re-exporting the shared types, so existing importers are unaffected.

The experiment card renders `MetricTrend` directly, since the card
already shows the experiment's name and description.
